Move API call code
